# frontend/lib/api/openapi.json
{
 "components": {
  "schemas": {
   "ActionView": {
    "properties": {
     "decided_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Decided At"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "expires_at": {
      "format": "date-time",
      "title": "Expires At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "enum": [
       "create_scenario",
       "run_optimization",
       "generate_report"
      ],
      "title": "Kind",
      "type": "string"
     },
     "message_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message Id"
     },
     "payload": {
      "additionalProperties": true,
      "title": "Payload",
      "type": "object"
     },
     "result": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Result"
     },
     "status": {
      "enum": [
       "pending",
       "executed",
       "rejected",
       "expired",
       "failed"
      ],
      "title": "Status",
      "type": "string"
     },
     "summary": {
      "title": "Summary",
      "type": "string"
     }
    },
    "required": [
     "id",
     "message_id",
     "kind",
     "summary",
     "status",
     "payload",
     "result",
     "error",
     "expires_at",
     "decided_at"
    ],
    "title": "ActionView",
    "type": "object"
   },
...
   "ComparisonNarrativeRequest": {
    "properties": {
     "baseline_run_id": {
      "title": "Baseline Run Id",
      "type": "string"
     },
     "locale": {
      "default": "fr",
      "enum": [
       "fr",
       "en"
      ],
      "title": "Locale",
      "type": "string"
     },
     "run_ids": {
      "items": {
       "type": "string"
      },
      "maxItems": 3,
      "minItems": 1,
      "title": "Run Ids",
      "type": "array"
     }
    },
    "required": [
     "baseline_run_id",
     "run_ids"
    ],
    "title": "ComparisonNarrativeRequest",
    "type": "object"
   },
...
   "ConversationCreate": {
    "properties": {
     "context": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PageContext"
       },
       {
        "type": "null"
       }
      ]
     },
     "title": {
      "anyOf": [
       {
        "maxLength": 160,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "title": "ConversationCreate",
    "type": "object"
   },
   "ConversationDetail": {
    "properties": {
     "actions": {
      "items": {
       "$ref": "#/components/schemas/ActionView"
      },
      "title": "Actions",
      "type": "array"
     },
     "context": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PageContext"
       },
       {
        "type": "null"
       }
      ]
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "messages": {
      "items": {
       "$ref": "#/components/schemas/MessageView"
      },
      "title": "Messages",
      "type": "array"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "format": "date-time",
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title",
     "created_at",
     "updated_at",
     "context",
     "messages",
     "actions"
    ],
    "title": "ConversationDetail",
    "type": "object"
   },
   "ConversationSummary": {
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "format": "date-time",
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "title",
     "created_at",
     "updated_at"
    ],
    "title": "ConversationSummary",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "MessageCreate": {
    "properties": {
     "content": {
      "maxLength": 4000,
      "minLength": 1,
      "title": "Content",
      "type": "string"
     },
     "context": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/PageContext"
       },
       {
        "type": "null"
       }
      ],
      "description": "What the user has on screen; default: the conversation's last context"
     }
    },
    "required": [
     "content"
    ],
    "title": "MessageCreate",
    "type": "object"
   },
...
   "NarrativeRequest": {
    "properties": {
     "locale": {
      "default": "fr",
      "enum": [
       "fr",
       "en"
      ],
      "title": "Locale",
      "type": "string"
     }
    },
    "title": "NarrativeRequest",
    "type": "object"
   },
   "NarrativeResponse": {
    "properties": {
     "locale": {
      "enum": [
       "fr",
       "en"
      ],
      "title": "Locale",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "prompt": {
      "title": "Prompt",
      "type": "string"
     },
     "text": {
      "description": "Numbers rendered by the backend; \u27e6?\u2026\u27e7 marks anything unverified",
      "title": "Text",
      "type": "string"
     },
     "verification": {
      "additionalProperties": true,
      "title": "Verification",
      "type": "object"
     }
    },
    "required": [
     "text",
     "verification",
     "locale",
     "model",
     "prompt"
    ],
    "title": "NarrativeResponse",
    "type": "object"
   },
...
   "PageContext": {
    "description": "What the user has open when asking (sent by the frontend, validated, never trusted blindly).",
    "properties": {
     "compare_run_ids": {
      "items": {
       "type": "string"
      },
      "maxItems": 4,
      "title": "Compare Run Ids",
      "type": "array"
     },
     "dataset_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dataset Id"
     },
     "page": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Page"
     },
     "run_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Run Id"
     },
     "scenario_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Scenario Id"
     }
    },
    "title": "PageContext",
    "type": "object"
   },
   "Page_ConversationSummary_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ConversationSummary"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[ConversationSummary]",
    "type": "object"
   },
...
   "ParseRequest": {
    "properties": {
     "dataset_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dataset Id"
     },
     "scenario_id": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Parse against this scenario's effective data",
      "title": "Scenario Id"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "ParseRequest",
    "type": "object"
   },
   "ParseResult": {
    "properties": {
     "changes": {
      "items": {
       "$ref": "#/components/schemas/ProposedChange"
      },
      "title": "Changes",
      "type": "array"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "prompt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Prompt"
     },
     "questions": {
      "items": {
       "type": "string"
      },
      "title": "Questions",
      "type": "array"
     },
     "rejected": {
      "items": {
       "$ref": "#/components/schemas/RejectedChange"
      },
      "title": "Rejected",
      "type": "array"
     }
    },
    "required": [
     "changes",
     "rejected",
     "questions"
    ],
    "title": "ParseResult",
    "type": "object"
   },
...
   "RejectedChange": {
    "properties": {
     "change": {
      "additionalProperties": true,
      "title": "Change",
      "type": "object"
     },
     "reason": {
      "title": "Reason",
      "type": "string"
     }
    },
    "required": [
     "change",
     "reason"
    ],
    "title": "RejectedChange",
    "type": "object"
   },
...
   "TurnResponse": {
    "properties": {
     "actions": {
      "items": {
       "$ref": "#/components/schemas/ActionView"
      },
      "title": "Actions",
      "type": "array"
     },
     "message": {
      "$ref": "#/components/schemas/MessageView"
     },
     "user_message": {
      "$ref": "#/components/schemas/MessageView"
     }
    },
    "required": [
     "user_message",
     "message",
     "actions"
    ],
    "title": "TurnResponse",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Harvest2Value API",
  "version": "2.0.0-dev"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v2/comparisons/narrative": {
   "post": {
    "operationId": "comparison_narrative_endpoint_api_v2_comparisons_narrative_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ComparisonNarrativeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NarrativeResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Run not found"
     },
     "422": {
      "description": "INCOMPARABLE"
     },
     "429": {
      "description": "RATE_LIMITED"
     },
     "502": {
      "description": "LLM_UPSTREAM_ERROR"
     },
     "503": {
      "description": "LLM_NOT_CONFIGURED: AI disabled"
     },
     "504": {
      "description": "LLM_TIMEOUT"
     }
    },
    "summary": "Comparison Narrative Endpoint",
    "tags": [
     "copilot"
    ]
   }
  },
  "/api/v2/copilot/actions/{action_id}/confirm": {
   "post": {
    "description": "Executes a pending proposal exactly once (idempotent: confirming again returns the same\nresult). A proposed run is queued like POST /runs.",
    "operationId": "confirm_action_api_v2_copilot_actions__action_id__confirm_post",
    "parameters": [
     {
      "in": "path",
      "name": "action_id",
      "required": true,
      "schema": {
       "title": "Action Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActionView"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Action not found"
     },
     "409": {
      "description": "ACTION_EXPIRED or ACTION_ALREADY_DECIDED"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Confirm Action",
    "tags": [
     "copilot"
    ]
   }
  },
  "/api/v2/copilot/actions/{action_id}/reject": {
   "post": {
    "operationId": "reject_action_api_v2_copilot_actions__action_id__reject_post",
    "parameters": [
     {
      "in": "path",
      "name": "action_id",
      "required": true,
      "schema": {
       "title": "Action Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActionView"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Action not found"
     },
     "409": {
      "description": "ACTION_EXPIRED or ACTION_ALREADY_DECIDED"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Reject Action",
    "tags": [
     "copilot"
    ]
   }
  },
  "/api/v2/copilot/conversations": {
   "get": {
    "operationId": "list_conversations_api_v2_copilot_conversations_get",
    "parameters": [
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ConversationSummary_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Conversations",
    "tags": [
     "copilot"
    ]
   },
   "post": {
    "operationId": "create_conversation_api_v2_copilot_conversations_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/ConversationCreate"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Conversation",
    "tags": [
     "copilot"
    ]
   }
  },
  "/api/v2/copilot/conversations/{conversation_id}": {
   "delete": {
    "operationId": "delete_conversation_api_v2_copilot_conversations__conversation_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Conversation",
    "tags": [
     "copilot"
    ]
   },
   "get": {
    "operationId": "get_conversation_api_v2_copilot_conversations__conversation_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Conversation",
    "tags": [
     "copilot"
    ]
   }
  },
  "/api/v2/copilot/conversations/{conversation_id}/messages": {
   "post": {
    "description": "One copilot turn. SSE events: `user_message`, `tool_call`, `tool_result`, `answer`\n(final message + proposed actions) or `error`, then `done`. Numbers in the answer are rendered\nby the backend from tool results; anything unverified is marked \u27e6?\u2026\u27e7. Proposals are only\n*pending* actions: nothing changes until POST /copilot/actions/{id}/confirm.",
    "operationId": "post_message_api_v2_copilot_conversations__conversation_id__messages_post",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "title": "Conversation Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "stream",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Stream"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MessageCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TurnResponse"
        }
       },
       "text/event-stream": {}
      },
      "description": "JSON, or SSE with `stream=true` / Accept: text/event-stream"
     },
     "404": {
      "description": "Conversation not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     },
     "429": {
      "description": "RATE_LIMITED"
     },
     "502": {
      "description": "LLM_UPSTREAM_ERROR"
     },
     "503": {
      "description": "LLM_NOT_CONFIGURED: AI disabled"
     },
     "504": {
      "description": "LLM_TIMEOUT"
     }
    },
    "summary": "Post Message",
    "tags": [
     "copilot"
    ]
   }
  },
...
  "/api/v2/runs/{run_id}/explanation/narrative": {
   "post": {
    "operationId": "explanation_narrative_api_v2_runs__run_id__explanation_narrative_post",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/NarrativeRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NarrativeResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Run not found"
     },
     "409": {
      "description": "Run has no plan"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     },
     "429": {
      "description": "RATE_LIMITED"
     },
     "502": {
      "description": "LLM_UPSTREAM_ERROR"
     },
     "503": {
      "description": "LLM_NOT_CONFIGURED: AI disabled"
     },
     "504": {
      "description": "LLM_TIMEOUT"
     }
    },
    "summary": "Explanation Narrative",
    "tags": [
     "copilot"
    ]
   }
  },
...
  "/api/v2/scenarios/parse": {
   "post": {
    "description": "Free text -> proposed typed changes, each validated and grounded in the sentence. Nothing is\nsaved: the user adds the changes in Scenario Studio. A plain question yields no change.",
    "operationId": "parse_scenario_api_v2_scenarios_parse_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ParseRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ParseResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "description": "Dataset or scenario not found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     },
     "429": {
      "description": "RATE_LIMITED"
     },
     "502": {
      "description": "LLM_UPSTREAM_ERROR"
     },
     "503": {
      "description": "LLM_NOT_CONFIGURED: AI disabled"
     },
     "504": {
      "description": "LLM_TIMEOUT"
     }
    },
    "summary": "Parse Scenario",
    "tags": [
     "copilot"
    ]
   }
  },
...
 }
}